refactor: move file_edit_history migration helper to files adapters layer
Move `migrate_file_history_unique_index` from the cross-cutting
`app/core/database_utils.py` to the file-domain adapters layer at
`app/files/adapters/migrations.py`, per `docs/ARCHITECTURE.md` §4.3
which requires domain-specific DDL to live under
`app/<domain>/adapters/migrations.py`.

The helper itself is unchanged — only its module location and import
paths move. The startup call ordering in `app/main.py` (immediately
after `Base.metadata.create_all`) is preserved.

Resolves #268
